refactor: move design primitives to shared
