+++ src/components/ide/StatusBar.tsx
@@ -1,7 +1,7 @@
 import { ArrowDown, ArrowUp, GitBranch } from 'lucide-react';
 import * as React from 'react';
-import { cn } from '@/components/lib/utils';
-import { Indicator } from '@/components/ui/Indicator';
+import { cn } from '@/shared/lib/utils';
+import { Indicator } from '@/shared/ui/Indicator';
 
 export interface StatusBarProps extends React.HTMLAttributes<HTMLDivElement> {
   branch?: string;
