+++ src/pages/dead-code/widgets/DeadCodeExplorer/ui/DeadCodeCategory.tsx
@@ -8,7 +8,6 @@ import { ChevronDown, ChevronRight } from 'lucide-react';
 import { useMemo } from 'react';
 import { activeActivityPageIdAtom } from '@/app/model/activityPageAtoms';
 import { FileTreeItem } from '@/components/ide/FileTreeItem.tsx';
-import { Checkbox } from '@/components/ui/Checkbox.tsx';
 import { filesAtom, viewModeAtom } from '@/entities/AppView/model/atoms';
 import { FileIcon } from '@/entities/SourceFileNode/ui/FileIcon.tsx';
 import { buildDeadCodeTree } from '@/features/Code/CodeAnalyzer/DeadCodeAnalyzer/lib/buildDeadCodeTree.ts';
@@ -23,6 +22,7 @@ import { useDeadCodeSelection } from '@/features/Code/CodeAnalyzer/DeadCodeSelec
 import { CategoryCheckbox } from '@/features/Code/CodeAnalyzer/DeadCodeSelection/ui/CategoryCheckbox.tsx';
 import { targetLineAtom } from '@/features/File/Navigation/model/atoms.ts';
 import { useOpenFile } from '@/features/File/OpenFiles/lib/useOpenFile.ts';
+import { Checkbox } from '@/shared/ui/Checkbox';
 import { TreeView } from '@/shared/ui/TreeView/TreeView.tsx';
 import { DeadCodeFolderItem } from './DeadCodeFolderItem.tsx';
