+++ src/pages/dead-code/widgets/DeadCodeExplorer/ui/DeadCodeFileItem.tsx
@@ -6,13 +6,13 @@ import { useSetAtom } from 'jotai';
 import React from 'react';
 import { activeActivityPageIdAtom } from '@/app/model/activityPageAtoms';
 import { FileTreeItem } from '@/components/ide/FileTreeItem.tsx';
-import { Checkbox } from '@/components/ui/Checkbox.tsx';
 import { viewModeAtom } from '@/entities/AppView/model/atoms';
 import { FileIcon } from '@/entities/SourceFileNode/ui/FileIcon.tsx';
 import type { DeadCodeItem } from '@/features/Code/CodeAnalyzer/DeadCodeAnalyzer/lib/deadCodeAnalyzer.ts';
 import { useDeadCodeSelection } from '@/features/Code/CodeAnalyzer/DeadCodeSelection/lib/useDeadCodeSelection.ts';
 import { targetLineAtom } from '@/features/File/Navigation/model/atoms.ts';
 import { useOpenFile } from '@/features/File/OpenFiles/lib/useOpenFile.ts';
+import { Checkbox } from '@/shared/ui/Checkbox';
 
 export const DeadCodeFileItem = React.forwardRef<
   HTMLDivElement,
