+++ src/pages/json-explorer/widgets/JsonExplorer/ui/SearchBar.tsx
@@ -3,7 +3,7 @@
  */
 
 import { Search } from 'lucide-react';
-import { Input } from '@/components/ui/Input';
+import { Input } from '@/shared/ui/Input';
 
 interface SearchBarProps {
   value: string;
