+++ src/shared/ui/Card.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { cn } from '@/components/lib/utils';
+import { cn } from '@/shared/lib/utils';
 
 const Card = React.forwardRef<HTMLDivElement, React.HTMLAttributes<HTMLDivElement> & { active?: boolean }>(
   ({ className, active, ...props }, ref) => (
