+++ src/shared/ui/Command.tsx
@@ -5,8 +5,8 @@ import { Command as CommandPrimitive } from 'cmdk';
 import { Search } from 'lucide-react';
 import * as React from 'react';
 
-import { cn } from '@/components/lib/utils';
-import { Dialog, DialogContent } from '@/components/ui/Dialog';
+import { cn } from '@/shared/lib/utils';
+import { Dialog, DialogContent } from '@/shared/ui/Dialog';
 
 const Command = React.forwardRef<
   React.ElementRef<typeof CommandPrimitive>,
