+++ src/shared/ui/Dialog.tsx
@@ -2,7 +2,7 @@ import * as DialogPrimitive from '@radix-ui/react-dialog';
 import { X } from 'lucide-react';
 import * as React from 'react';
 
-import { cn } from '@/components/lib/utils';
+import { cn } from '@/shared/lib/utils';
 
 const Dialog = DialogPrimitive.Root;
