+++ src/shared/ui/Indicator.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { cn } from '@/components/lib/utils';
+import { cn } from '@/shared/lib/utils';
 
 export interface IndicatorProps extends React.HTMLAttributes<HTMLDivElement> {
   variant?: 'inactive' | 'active' | 'success' | 'warning' | 'error' | 'working';
